% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{bayes.factors}
\alias{bayes.factors}
\title{Calculate Bayes factors and posterior model probabilities}
\usage{
bayes.factors(...)
}
\arguments{
\item{...}{list of marginal likelihood objects, see details}
}
\value{
A list with elements \code{bf}, the Bayes factors; and \code{pr}, the
posterior model probabilities.
}
\description{
Calculate Bayes factors and posterior model probabilities
}
\details{
Input is a list of marginal likelihood objects, with each object generated
 by either \code{stepping.stones()} or \code{gauss.quad()}.
}
\author{
Mario dos Reis
}
