% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{gauss.quad}
\alias{gauss.quad}
\title{Estimate marginal likelihood by thermodynamic integration}
\usage{
gauss.quad(mcmcf = "mcmc.txt", betaf = "beta.txt")
}
\arguments{
\item{mcmcf}{character, mcmc output file name}

\item{betaf}{character, file with beta values}
}
\value{
A list with components \code{logml}, the log-marginal likelihood estimate;
\code{se}, the standard error of the estimate; \code{mean.logl}, the mean of
log-likelihood values sampled for each beta; and \code{b}, the beta values
used.
}
\description{
Estimate marginal likelihood by thermodynamic integration and Gauss-Legendre
quadrature from a sample of \code{n} power posterior MCMC chains sampled
with mcmctree (or bpp).
}
\details{
The MCMC samples should be stored in a directory structure created
by \code{make.bfctlf} with \code{method = "gauss-quad"}. The function
will read the stored log-likelihood values and calculate the log-marginal
likelihood.

Numerical integration is done using Gauss-Legendre quadrature. See Rannala
and Yang (2017) for details (also dos Reis et al. 2017, Appendix 2).
}
\references{
Rannala B and Yang Z. (2017) Efficient Bayesian species tree inference under
the multispecies coalescent. \emph{Systematic Biology} 66: 823-842.

dos Reis et al. (2017) Using phylogenomic data to explore the effects of
relaxed clocks and calibration strategies on divergence time estimation:
Primates as a test case. \emph{bioRxiv}
}
\seealso{
\code{\link{make.bfctlf}} to prepare directories and mcmctree or bpp control files
to calculate the power posterior.
}
\author{
Mario dos Reis
}
