% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{make.beta}
\alias{make.beta}
\title{Make beta values for marginal likelihood calculation}
\usage{
make.beta(n, method = c("step-stones", "gauss-quad"), a = 5)
}
\arguments{
\item{n}{numeric, number of beta points}

\item{method}{character, the method to choose the beta points, see details}

\item{a}{numeric, exponent for stepping stones beta generation, see details}
}
\value{
Numeric vector with \code{n} beta values
}
\description{
Make appropriate beta values
}
\details{
If \code{method = "step-stones"}, the beta values are given by the
formula

\deqn{\beta_{i}=\left(\frac{i-1}{n}\right)^{a}.}

Values of \code{a} between 5 to 8 appear appropriate. Large \code{a} values
produce beta values close to zero.

If \code{method = "gauss-quad"}, the beta values are calculated according to
the \code{n} Gauss-Legendre quadrature rule (see Rannala and Yang, 2017).
}
\references{
Rannala B and Yang Z (2017) Efficient Bayesian species tree inference under
the multispecies coalescent. \emph{Systematic Biology}, 66: 823--842.
}
\seealso{
The generated beta values are suitable input for \code{\link{make.bfctlf}}.
}
\author{
Mario dos Reis
}
