% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{make.bfctlf}
\alias{make.bfctlf}
\title{Prepare mcmctree or bpp control files for marginal likelihood calculation}
\usage{
make.bfctlf(beta, ctlf = "bpp.ctl", betaf = "beta.txt")
}
\arguments{
\item{beta}{numeric vector of beta values}

\item{ctlf}{character, mcmctree or bpp control file template}

\item{betaf}{character, file onto which to write selected beta values}
}
\description{
Prepare mcmctree or bpp control files for marginal likelihood calculation
}
\details{
This function generates a set of \code{n} directories each
containing a modified \code{ctlf} control file with the appropriate beta
value to run mcmctree (or bpp) to obtain MCMC samples under the required
power-posterior distribution. For the general theory of marginal likelihood
calculation with power posteriors see Yang (2014).

The beta values are printed to \code{betaf}.
}
\references{
Yang Z (2014) \emph{Molecular Evolution: A Statistical Approach}. Oxford
University Press. Pages 256--260.
}
\seealso{
\code{\link{make.beta}}, \code{\link{stepping.stones}}
}
\author{
Mario dos Reis
}
