% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc2time.R
\docType{data}
\name{microcebus}
\alias{microcebus}
\title{A BPP A00 MCMC sample for a mouse lemur phylogeny}
\format{An object of class \code{list} of length 2.}
\usage{
microcebus
}
\description{
A BPP A00 MCMC sample for a mouse lemur phylogeny
}
\keyword{datasets}
