% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc2time.R
\name{msc2time.r}
\alias{msc2time.r}
\title{Calibrate a BPP A00 analysis using a prior on the rate}
\usage{
msc2time.r(mcmc, u.mean, u.sd, g.mean, g.sd)
}
\description{
Calibrate a BPP A00 analysis using a prior on the rate
}
\author{
Mario dos Reis
}
