% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc2time.R
\name{msc2time.t}
\alias{msc2time.t}
\title{Calibrate a BPP A00 analysis}
\usage{
msc2time.t(mcmc, node.name, bounds)
}
\arguments{
\item{mcmc}{A data frame containing the MCMC output of a BPP A00 analysis}

\item{node.name}{A character vector of length one with the name of the node to which the calibration will be applied}

\item{bounds}{A numeric vector of length two with age bounds for the calibration}
}
\value{
A data frame with the calibrated times and molecular rate.
}
\description{
Calibrate a multi-species coalescent phylogeny to
absolute time using a fossil calibration.
}
\details{
The function will obtain a sample of times from a uniform distribution
delimited by `bounds`. The sampled times are then used to
calculate the molecular rate, and then re-scale the relative
times (tau's) for the other nodes in `mcmc`.
}
\examples{
data(hominids)

# Calibrate the hominid phylogeny with a fossil calibration of
# between 6.5 to 10 Ma for the human-chimp divergence.
calmsc <- msc2time.t(mcmc=hominids$mcmc, node="7humanchimp", bounds=c(6.5, 10))

# posterior age of human-chimp (this is the calibration)
plot(density(calmsc$t_7humanchimp, adj=.1), xlab="Time (Ma)")
rug(calmsc$t_7humanchimp)

# posterior age of human-goriall (root of the tree)
plot(density(calmsc$t_5humanchimpgorillaorang, adj=.1), xlab="Time (Ma)")
rug(calmsc$t_5humanchimpgorillaorang)

\dontrun{
# if you have the coda package installed, you can get the
# HPD intervals
coda::HPDinterval(coda::as.mcmc(calmsc))
}

}
\author{
Mario dos Reis
}
