% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShiftedLognormal.R
\name{ShiftedLognormal}
\alias{ShiftedLognormal}
\alias{dslnorm}
\alias{pslnorm}
\alias{qslnorm}
\alias{rslnorm}
\title{The Shifted Log-normal Distribution}
\usage{
dslnorm(x, shift, meanlog = 0, sdlog = 1, log = FALSE)

pslnorm(q, shift, meanlog = 0, sdlog = 1, lower.tail = TRUE,
  log.p = FALSE)

qslnorm(p, shift, meanlog = 0, sdlog = 1, lower.tail = TRUE,
  log.p = FALSE)

rslnorm(n, shift, meanlog = 0, sdlog = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{shift}{vector of shifts.}

\item{meanlog, sdlog}{mean and standard deviation of the distribution on the
log scale with default values of 0 and 1 respectively.}

\item{log, log.p}{logical; if TRUE, probabilities p are given as log(p).}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \leq
x]}, otherwise, \eqn{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of observations. If \code{length(n) > 1}, the length is taken
to be the number required.}
}
\value{
\code{dslnorm} gives the density, \code{pslnorm} gives the distribution
  function, \code{qslnorm} gives the quantile function, and \code{rslnorm}
  generates random deviates.

  The length of the result is determined by \code{n} for \code{rlnorm}, and
  is the maximum of the lengths of the numerical arguments for the other
  functions.

  The numerical arguments other than \code{n} are recycled to the length of
  the result. Only the first elements of the logical arguments are used.
}
\description{
Density, distribution and quantile functions, and random number generation
for the shifted log-normal distribution.
}
\details{
Let \eqn{Y} have a log-normal distribution with parameters \eqn{\mu}
  (\code{meanlog}) and \eqn{\sigma} (\code{sdlog}). Then \eqn{X = Y + s} has
  a shifted log-normal distribution with shift \eqn{s} (\code{shift}), mean
  \eqn{E(X) = exp(\mu + 1/2 \sigma^2) + s} and variance \eqn{Var(X) =
  exp(2*\mu + \sigma^2)*(exp(\sigma^2) - 1)}.

  Note \code{[dpqr]slnorm} are wrappers for the corresponding
  \code{[dpqr]lnorm} functions.
}
\examples{
curve(dslnorm(x, shift=6.5), from=0, to=15, n=1e3)
rr <- rslnorm(1e3, shift=6.5)
lines(density(rr, adj=.1), lty=2)

all.equal (qslnorm(c(.025, .9), shift=6.5) - 6.5, qlnorm(c(.025, .9)))
all.equal (pslnorm(10, shift=6.5), plnorm(10 - 6.5))

}
\seealso{
\code{\link{Lognormal}}
}
