% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{stepping.stones}
\alias{stepping.stones}
\title{Estimate marginal likelihood by stepping stones}
\usage{
stepping.stones(mcmcf = "mcmc.txt", betaf = "beta.txt")
}
\arguments{
\item{mcmcf}{character, mcmc output file name}

\item{betaf}{character, file with beta values}
}
\value{
A list with components \code{logml}, the log-marginal likelihood estimate;
\code{se}, the standard error of the estimate; \code{mean.logl}, the mean of
log-likelihood values sampled for each beta; and \code{b}, the beta values
used.
}
\description{
Estimate the marginal likelihood using the stepping stones
method from a sample of \code{n} power posterior MCMC chains sampled
with mcmctree (or bpp).
}
\details{
The MCMC samples should be stored in a directory structure created
by \code{make.bfctlf} with \code{method = "step-stones"}. The function will
read the stored log-likelihood values and calculate the log-marginal
likelihood.

An approximation based on the Delta method is used to calculate the standard
error (see Xie et al. 2011). Warnings are given if the approximation appears
unreliable.
}
\references{
Xie et al. (2011) Improving marginal likelihood estimation for Bayesian
phylogenetic model selection. \emph{Systematic Biology}, 60: 150--160.
}
\seealso{
\code{\link{make.bfctlf}} to prepare directories and mcmctree or bpp control
files to calculate the power posterior.
}
\author{
Mario dos Reis
}
