% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-lhd.R
\name{bayes.factors}
\alias{bayes.factors}
\title{Calculate Bayes factors and posterior model probabilities}
\usage{
bayes.factors(..., prior = NULL, boot = TRUE, n = 1000, prob = 0.95)
}
\arguments{
\item{...}{list of marginal likelihood objects, see details}

\item{prior}{numeric, the prior model probabilities}

\item{boot}{logical, whether to perform parametric boostrap of probabilities}

\item{n}{numeric, number of bootstrap samples}

\item{prob}{numeric, the probability used to calculate the boostrap CI}
}
\value{
A list with elements \code{bf}, the Bayes factors; \code{pr}, the posterior
model probabilities; \code{prior} the prior model probabilities and, if
\code{boot = TRUE}, \code{pr.ci} the equal-tail bootstrap confidence interval.
}
\description{
Calculate Bayes factors and posterior model probabilities
}
\details{
Input is a list of marginal likelihood objects, with each object generated by
either \code{stepping.stones()} or \code{gauss.quad()}. If \code{boot =
TRUE}, parametric bootstrap is performed by assuming the log-marginal
likelihood estimates are normally distributed with standard deviation equal
to the standard error. The re-sampled \code{n} marginal log-likelihoods are
used to estimate re-sampled posterior probabilities, and to calculate an
equal-tail bootstrap confidence interval for these.

Note that the length of \code{prior} should be the same as the number of
models being compared. The \code{prior} is rescaled so that
\code{sum(prior) == 1}.
}
\examples{
# See Table 5 in dos Reis et al. (2018, Syst. Biol., 67: 594-615)
# Bayesian selection of relaxed clock models for the 1st and 2nd sites
# of mitochondrial protein-coding genes of primates
# Models: strick clock, independent-rates, and autocorrelated-rates
sc <- list(); sc$logml <- -16519.03; sc$se <- .01
ir <- list(); ir$logml <- -16480.58; ir$se <- .063
ar <- list(); ar$logml <- -16477.82; ar$se <- .035
bayes.factors(sc, ir, ar)
bayes.factors(sc, ir, ar, prior=c(.25,.5,.25))
bayes.factors(sc, ir, ar, prior=c(0,1,0))

}
\author{
Mario dos Reis
}
