% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msc2time.R
\docType{data}
\name{hominids}
\alias{hominids}
\title{A BPP A00 MCMC sample for an hominid phylogeny}
\format{An object of class \code{list} of length 2.}
\usage{
hominids
}
\description{
A BPP A00 MCMC sample for an hominid phylogeny
}
\keyword{datasets}
